requirements.txt
streamlit 
langchain 
langchain_openai 
langchain_postgres 
psycopg2 
streamlit_oauth 
python-dotenv
